requirements.txt
streamlit==1.42.2
Pillow==10.4.0
yfinance==0.2.54
ta==0.11.0
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
seaborn==0.13.2
japanize_matplotlib==1.1.3
matplotlib==3.7.5
plotly==5.24.1
requests==2.32.3
bs4==0.0.2
lxml==5.2.1
html5lib==1.1
lightgbm==4.6.0
mplfinance==0.12.10b0
openpyxl==3.1.5
streamlit-aggrid==1.1.2